+++ src/util.ts
@@ -1,5 +1,5 @@
-import { exec } from "child_process";
-import { promisify } from "util";
+import { exec } from "node:child_process";
+import { promisify } from "node:util";
 const execAsPromised = promisify(exec);
 
 import { error, info, setFailed } from "@actions/core";
